move tests for subscribersPool
